requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
openai==0.28.1
python-dotenv==1.0.0
PyJWT==2.8.0
Flask-SocketIO==5.3.6
python-socketio==5.9.0
redis==5.0.1
Pillow==10.0.0

